Add admin and editor routes for user and metadata management
- Introduced new admin blueprint with routes for user management, including adding, editing, and deleting users.
- Added endpoints for managing metadata, including ranks and organizations, with appropriate permission checks.
- Implemented audit logging for user and metadata changes to enhance traceability.
- Registered new routes in app.py to integrate with the existing application structure.
- Enhanced error handling and user feedback for CRUD operations in user management and metadata management.
