+++ templates/location_form.html
@@ -144,7 +144,7 @@
                         </div>
 
                         <div class="d-flex justify-content-between">
-                            <a href="{{ url_for('main.locations_list') }}" class="btn btn-secondary">
+                            <a href="{{ url_for('locations.locations_list') }}" class="btn btn-secondary">
                                 <i class="fas fa-arrow-left me-1"></i>Back to Locations
                             </a>
                             <button type="submit" class="btn btn-primary">
